Ensure _loaderHelpers registers actual filenames
Slicing a path by the last 3 characters works for two character
extensions, e.g. `'.js'` or `'.cs'` but not for longer extensions
(`'.es6'`, `'.javascript'`, `'.coffeescript'` for example).

Using `Path.extname` will extract the proper extension from a path,
which can then be used to determine the length of the extension proper,
and thereby allowing helpers with longer extension names.