Make all rendering errors available to 'onPostHandler'.
Currently, rendering happens right before the response is sent to
the client. This means that errors and other rendering results are
not available to 'onPostHandler' and 'onPreResponse' handlers. This
change moves the rendering step to an earlier point in the request
lifecycle to allow the results to be processed after the handler
has completed.

Fixes #10.