Update plugin attributes.
In some use cases it may make sense to load vision before any
connections have been configured. Similarly, multiple plugins in an
application may try to load vision. Since vision only deals with
template rendering, it is possible to do both of these if the
'connections' and 'once' attributes are provided in the plugin
attributes.

Resolves #63.